Move texture hydration for serialized projects to the editor package so it can be reused there
