+++ packages/demo/src/configuration/smartFilters/hardCoded/videoWithFrame.ts
@@ -2,10 +2,10 @@ import type { ThinEngine } from "@babylonjs/core/Engines/thinEngine";
 import { ConnectionPointType, InputBlock, SmartFilter, createImageTexture } from "@babylonjs/smart-filters";
 import { BlackAndWhiteBlock } from "../../blocks/effects/blackAndWhiteBlock";
 import { HardCodedSmartFilterNames } from "./hardCodedSmartFilterNames";
-import { createVideoTextureAsync } from "../../../helpers/createVideoTexture";
 import { BlurBlock } from "../../blocks/effects/blurBlock";
 import { FrameBlock } from "../../blocks/effects/frameBlock";
 import type { SmartFilterRenderer } from "../../../smartFilterRenderer";
+import { createVideoTextureAsync } from "@babylonjs/smart-filters-editor";
 
 export async function createVideoWithFrameSmartFilter(
     engine: ThinEngine,
