+++ packages/editor/src/graphSystem/getEditorData.ts
@@ -18,6 +18,7 @@ export function getTextureInputBlockEditorData(
             anisotropicFilteringLevel: internalTexture?.anisotropicFilteringLevel ?? null,
             flipY: internalTexture?.invertY ?? true,
             forcedExtension: internalTexture?._extension ?? null,
+            dispose: null,
         };
     }
